+++ packages/pigeon/mock_handler_tester/test/widget_test.dart
@@ -14,10 +14,32 @@ class Mock implements TestHostApi {
   }
 }
 
+class MockNested implements TestNestedApi {
+  bool didCall = false;
+  @override
+  SearchReply search(Nested arg) {
+    didCall = true;
+    if (arg.request == null) {
+      return SearchReply();
+    } else {
+      return SearchReply()..result = arg.request.query;
+    }
+  }
+}
+
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
 
-  test('description', () async {
+  test('simple', () async {
+    final NestedApi api = NestedApi();
+    final MockNested mock = MockNested();
+    TestNestedApi.setup(mock);
+    final SearchReply reply = await api.search(Nested()..request = null);
+    expect(mock.didCall, true);
+    expect(reply.result, null);
+  });
+
+  test('nested', () async {
     final Api api = Api();
     final Mock mock = Mock();
     TestHostApi.setup(mock);